+++ package.json
@@ -58,7 +58,7 @@
     "redux-thunk": "^2.4.2",
     "stream": "^0.0.2",
     "tailwindcss": "^3.3.3",
-    "tlsn-js": "0.1.0-alpha.7.1",
+    "tlsn-js": "0.1.0-alpha.9",
     "tlsn-js-v5": "npm:tlsn-js@0.1.0-alpha.5.4",
     "ws": "^8.16.0"
   },
